+++ functions/function.c
@@ -16,7 +16,7 @@ void startFunction(char *name, enum Type type) {
         free(function_mode->body);
         function_mode->body = "";
         free(name);
-        for (int i = 0; i < function_parameters_mode->parameter_count; i++) {
+        for (unsigned short i = 0; i < function_parameters_mode->parameter_count; i++) {
             Symbol* parameter = function_parameters_mode->parameters[i];
             removeSymbol(parameter);
         }
@@ -49,7 +49,7 @@ void startFunction(char *name, enum Type type) {
     function_mode->decision_functions.capacity = 0;
     function_mode->decision_functions.size = 0;
 
-    int parent_context = 1;
+    unsigned short parent_context = 1;
     if (module_path_stack.size > 1) parent_context = 2;
     function_mode->context = malloc(1 + strlen(module_path_stack.arr[module_path_stack.size - parent_context]));
     strcpy(function_mode->context, module_path_stack.arr[module_path_stack.size - parent_context]);
@@ -85,7 +85,7 @@ void startFunction(char *name, enum Type type) {
     function_mode->parameter_count = function_parameters_mode->parameter_count;
 
 
-    for (int i = 0; i < function_mode->parameter_count; i++) {
+    for (unsigned short i = 0; i < function_mode->parameter_count; i++) {
         Symbol* parameter = function_mode->parameters[i];
         parameter->param_of = function_mode;
     }
@@ -128,7 +128,7 @@ void callFunction(char *name, char *module) {
     _Function* function = getFunction(name, module);
 
     scope_override = function;
-    for (int i = 0; i < function->parameter_count; i++) {
+    for (unsigned short i = 0; i < function->parameter_count; i++) {
         Symbol* parameter = function->parameters[i];
         Symbol* parameter_call = function_parameters_mode->parameters[i];
 
@@ -187,7 +187,7 @@ void callFunction(char *name, char *module) {
         return;
     }
 
-    for (int i = 0; i < function->parameter_count; i++) {
+    for (unsigned short i = 0; i < function->parameter_count; i++) {
         Symbol* parameter = function->parameters[i];
         removeSymbolByName(parameter->secondary_name);
     }
@@ -236,7 +236,7 @@ void printFunctionTable() {
         char *context_temp = malloc(1 + strlen(function->context));
         strcpy(context_temp, function->context);
         printf(
-            "\t{name: %s, type: %i, parameter_count: %i, decision_length: %i, context: %s, module_context: %s, module: %s} =>\n",
+            "\t{name: %s, type: %u, parameter_count: %hu, decision_length: %u, context: %s, module_context: %s, module: %s} =>\n",
             function->name,
             function->type,
             function->parameter_count,
@@ -297,7 +297,7 @@ void addFunctionCallParameterBool(bool b) {
     addSymbolToFunctionParameters(symbol);
 }
 
-void addFunctionCallParameterInt(int i) {
+void addFunctionCallParameterInt(long long i) {
     union Value value;
     value.i = i;
     Symbol* symbol = addSymbol(NULL, K_NUMBER, value, V_INT);
@@ -384,7 +384,7 @@ void freeFunction(_Function* function) {
     free(function->body);
     free(function->name);
     free(function->parameters);
-    for (int i = 0; i < function->decision_functions.size; i++) {
+    for (unsigned i = 0; i < function->decision_functions.size; i++) {
         free(function->decision_expressions.arr[i]);
         free(function->decision_functions.arr[i]);
     }
@@ -465,7 +465,7 @@ void executeDecision(_Function* function) {
     strcpy(name, __LANGUAGE_NAME__);
     Symbol* symbol = addSymbol(name, K_BOOL, value, V_BOOL);
 
-    for (int i = 0; i < function->decision_functions.size; i++) {
+    for (unsigned i = 0; i < function->decision_functions.size; i++) {
         expression_buffer = strcat_ext(expression_buffer, __LANGUAGE_NAME__);
         expression_buffer = strcat_ext(expression_buffer, " = ");
         expression_buffer = strcat_ext(expression_buffer, function->decision_expressions.arr[i]);
@@ -515,14 +515,14 @@ void addFunctionNameToFunctionNamesBuffer(char *name) {
 }
 
 void freeFunctionNamesBuffer() {
-    for (int i = 0; i < function_names_buffer.size; i++) {
+    for (unsigned i = 0; i < function_names_buffer.size; i++) {
         free(function_names_buffer.arr[i]);
     }
     function_names_buffer.size = 0;
 }
 
 bool isInFunctionNamesBuffer(char *name) {
-    for (int i = 0; i < function_names_buffer.size; i++) {
+    for (unsigned i = 0; i < function_names_buffer.size; i++) {
         if (strcmp(function_names_buffer.arr[i], name) == 0) return true;
     }
     return false;
